requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2  # Added for MongoDB SRV support